requirements.txt
Django>=2.0,<3.0
djangorestframework>=3.0.0
psycopg2-binary>=2.8